提交一波2018.8.15

+++ server/src/main/java/com/sh/wxa/module/activity/msg/ActivityIdRequest.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 public class ActivityIdRequest implements Message {
 
-    private int activityId;
+    private Long activityId;
 
 }
